[TASK] Move most development tools from PHIVE to Composer
We now only use PHIVE for those tools that are not compatible
with the PHP versions we currently support (7.2-7.4).
This currently is only the PHP copy-paste detector
(which requires PHP >= 7.3).

This also (as a nice side effect) provides `ddev composer normalize`
out of the box.

For YAML linting, we now use the `symfony/yaml` package (which already
was installed as an indirected dependency, and which we now explicitly
require).

Fixes #274
Fixes #279