[DOCS] List and describe all composer commands
The composer commands are the core value of this extension. List
all commands in the documentation and copy the descriptions from
the composer.json. This lets the global search at docs.typo3.org
find these commands.

Remove superfluous Linters page, which is integrated now in the
Running page.

Use sentence case in page titles, to conform to the TYPO3 documentation
standards.