[TASK] Move the coverage directories (#427)
There is no need to have the code coverage data directly generated by
the tests run in a non-hidden directory. So now the code coverage is
recorded within `.Build/coverage`.

Only the merged coverage data still needs to be located in `build/logs`
as Coveralls (which we're planning to use) relies on this.