[TASK] Add `.gitignore` entry for JetBrains Fleet editor (#642)
JetBrains created a new product named "Fleet" as
lightweight editor. This tool tends to write it's
config files to `.fleet/` folders, like all of the
IntellJ based IDE's like PHPStorm uses the `.idea`
folder.

This change adds this config folder to `.gitignore`
to avoid adding this folder to a patch if Fleet
is used to create a patch.