[FEATURE] Add automerging of green Dependabot PRs
Now dependency updates by Dependabot will be automerged if/when all
required checks are green.

This will reduce the repetetive work of merging all these update
PRs, and we have been heavily relying on our CI pipeline to check
if a depenceny update is safe to merge anyway.

Attaching any of the labels `wip` or `blocked` to a PR will block the
automerging.