[TASK] Unify web-dir, bin-dir
The following are changed:

- web-dir: ".Build/public" => ".Build/Web"
- bin-dir: not specified, defaulted to .Build/vendor/bin
  => ".Build/bin"

The vendor-dir was alreaday ".Build/vendor" and is left
unchanged.

Since the above paths may be used in scripts, composer.json,
build pipelines, configuration etc. it facilitates
contribution across various extensions (and the core) if
the same defaults are used.

The goal of patches related to #802 is to unify the paths
across TYPO3 extensions, specifically focusing on the
"official" extensions (such as tea, Documentation examples
extension etc.).

- .Build has already been well established (and is used by
"tea" as well.
- web-dir (".Build/Web") is currently not used consistently
  (across official extensions). Either ".Build/public", ".Build/Web"
  or ".Build/web" or "public" is used. We use ".Build/Web"
  now as that is already well established in "styleguide"
  and other extensions. (This choice is not better than
  the other, but is more commonly used and has led to a
  consensus in preliminary decision making process.)
- bin-dir: Here ".Build/bin" is mostly used (if defined at all)
- vendor-dir: Here ".Build/vendor" is mostly used (if defined
  at all)

Related: #802