[TASK] Make the development dependencies more specific (#828)
Even for a constraint like `^6.2`, we explicitly state the bug-fix version
`^6.2.0`. Even though this is semantically the same, this communicates to
the human reader that we explicitly allow version 6.2.0 instead of
"any 6.2 version".

Also bump `friendsofphp/php-cs-fixer`.

For dev dependencies that are also transitive dependencies of
`typo3/cms-core`, align our dependencies with the `typo3/cms-core`
dependencies. This allows us to catch bugs with the lowest dependencies
that would go unnoticed if we required a higher version as a dev depdency.