[TASK] Restructure the label keys
Label keys should be specific to their context. This allows having
different labels for the same thing in different contexts (e.g.,
having different headings for the tea title in the regular list view
and the FE editor).

We take the plugin as the context here, which allows us to reuse labels
between different actions of the same plugin (e.g., between the list view
and single view).