[TASK] Add PHPStan extension for cognitive load
A new composer package "tomasvotruba/cognitive-complexity" is added.
This one is a PHPStan extension which will check the cognitive
complexity of classes and functions/methods.

We have very low numbers as our examples are not too complex yet.
Real projects adopting the extension might configure higher values in
the beginning or use the baseline approach.

Resolves: #1154