[TASK] Stop storing development tool PHARs in the repository
Tools that are downloadable should in general be downloaded during
the CI runs, not stored in the repository.

As the only PHIVE PHAR we're using is PCOV (for code coverage, which
we usually only run on CI), we can skip providing PHIVE and rely on
CI to provide PHIVE for us. This leaves only PCOV to be installed
via PHIVE.

Fixes #267