[BUGFIX] Enhance container documentation rendering
The new php based rendering container has been added
recently to render the documentation. The image come
with a internal uid/gid switch to mitigate permission
issues with docker due to a missimplementation, which
podman not suffers by proper using the kernal namespacing.

This change ensures to create a folder before hand with
the correct permissions to mitigate this for docker. On
top of that, the external user set is added for the direct
invokiation and also the interactive switch to avoid failure
in CI context usages.

To avoid permission issues with previously created folder,
a `chown` command is added to ensure correct permission on
that folder.

Resolves: #1283