[BUGFIX] Refactor that pageUid stuff with assert and type of int
With the usage of an integer in functional test for the command
we get rid of the issue with asert is int or cast mixed variable to int.
Test are running fine locally. The page id should be just an integer
and nobody will be set that in '' right?

Related: #1120